Restore async deletion of external services

+++ cmd/frontend/graphqlbackend/external_services.go
@@ -189,8 +189,17 @@ func (r *schemaResolver) DeleteExternalService(ctx context.Context, args *delete
 		return nil, err
 	}
 
-	if err := r.db.ExternalServices().Delete(ctx, id); err != nil {
-		return nil, err
+	if args.Async {
+		// run deletion in the background and return right away
+		go func() {
+			if err := r.db.ExternalServices().Delete(ctx, id); err != nil {
+				r.logger.Error("Background external service deletion failed", log.Error(err))
+			}
+		}()
+	} else {
+		if err := r.db.ExternalServices().Delete(ctx, id); err != nil {
+			return nil, err
+		}
 	}
 
 	return &EmptyResponse{}, nil
